Remove AbstractXFormInstance
