+++ corehq/form_processor/models/abstract.py
@@ -1,131 +1,12 @@
 import collections
 from abc import ABCMeta, abstractmethod
 
-from casexml.apps.phone.exceptions import MissingSyncLog
-
 from memoized import memoized
-from couchforms import const
 
 from .mixin import CaseToXMLMixin
 
 
 DEFAULT_PARENT_IDENTIFIER = 'parent'
-
-
-class AbstractXFormInstance(object):
-
-    # @property
-    # def form_id(self):
-    #     raise NotImplementedError()
-
-    user_id = None
-
-    @property
-    def attachments(self):
-        """
-        Get the extra attachments for this form. This will not include
-        the form itself
-        """
-        raise NotImplementedError
-
-    @property
-    def form_data(self):
-        raise NotImplementedError()
-
-    @property
-    def metadata(self):
-        raise NotImplementedError()
-
-    @property
-    def is_normal(self):
-        raise NotImplementedError()
-
-    @property
-    def is_archived(self):
-        raise NotImplementedError()
-
-    @property
-    def is_deprecated(self):
-        raise NotImplementedError()
-
-    @property
-    def is_duplicate(self):
-        raise NotImplementedError()
-
-    @property
-    def is_error(self):
-        raise NotImplementedError()
-
-    @property
-    def is_submission_error_log(self):
-        raise NotImplementedError()
-
-    @property
-    def is_deleted(self):
-        raise NotImplementedError()
-
-    # @property
-    # def deletion_id(self):
-    #     raise NotImplementedError
-
-    def auth_context(self):
-        raise NotImplementedError()
-
-    def get_data(self, xpath):
-        raise NotImplementedError()
-
-    def get_attachment(self, attachment_name):
-        raise NotImplementedError()
-
-    def archive(self, user_id=None, trigger_signals=True):
-        raise NotImplementedError()
-
-    def unarchive(self, user_id=None):
-        raise NotImplementedError()
-
-    def get_xml_element(self):
-        raise NotImplementedError()
-
-    def get_xml(self):
-        raise NotImplementedError()
-
-    def save(self, *args, **kwargs):
-        raise NotImplementedError()
-
-    def set_submission_properties(self, submission_post):
-        raise NotImplementedError()
-
-    def soft_delete(self):
-        raise NotImplementedError()
-
-    def to_json(self):
-        raise NotImplementedError()
-
-    @classmethod
-    def get(self, xform_id):
-        raise NotImplementedError()
-
-    @property
-    def xml_md5(self):
-        raise NotImplementedError()
-
-    @property
-    def type(self):
-        return self.form_data.get(const.TAG_TYPE, "")
-
-    @property
-    def name(self):
-        return self.form_data.get(const.TAG_NAME, "")
-
-    @memoized
-    def get_sync_token(self):
-        from casexml.apps.phone.models import get_properly_wrapped_sync_log
-        if self.last_sync_token:
-            try:
-                return get_properly_wrapped_sync_log(self.last_sync_token)
-            except MissingSyncLog:
-                pass
-        return None
 
 
 def get_index_map(indices):
