+++ corehq/form_processor/models/forms.py
@@ -40,7 +40,6 @@ from ..exceptions import (
 from ..submission_process_tracker import unfinished_archive
 from ..system_action import system_action
 from ..track_related import TrackRelatedChanges
-from .abstract import AbstractXFormInstance
 from .attachment import AttachmentContent, AttachmentMixin
 from .mixin import SaveStateMixin
 from .util import attach_prefetch_models, sort_with_id_list
@@ -413,8 +412,8 @@ class XFormOperationManager(RequireDBManager):
         return list(self.partitioned_query(form_id).filter(form_id=form_id).order_by('date'))
 
 
-class XFormInstance(PartitionedModel, models.Model, RedisLockableMixIn, AttachmentMixin,
-                    AbstractXFormInstance, TrackRelatedChanges):
+class XFormInstance(PartitionedModel, models.Model, RedisLockableMixIn,
+                    AttachmentMixin, TrackRelatedChanges):
     partition_attr = 'form_id'
 
     # states should be powers of 2
@@ -615,6 +614,25 @@ class XFormInstance(PartitionedModel, models.Model, RedisLockableMixIn, Attachme
         from ..utils import clean_metadata
         if const.TAG_META in self.form_data:
             return XFormPhoneMetadata.wrap(clean_metadata(self.form_data[const.TAG_META]))
+
+    @property
+    def type(self):
+        return self.form_data.get(const.TAG_TYPE, "")
+
+    @property
+    def name(self):
+        return self.form_data.get(const.TAG_NAME, "")
+
+    @memoized
+    def get_sync_token(self):
+        from casexml.apps.phone.exceptions import MissingSyncLog
+        from casexml.apps.phone.models import get_properly_wrapped_sync_log
+        if self.last_sync_token:
+            try:
+                return get_properly_wrapped_sync_log(self.last_sync_token)
+            except MissingSyncLog:
+                pass
+        return None
 
     def soft_delete(self):
         type(self).objects.soft_delete_forms(self.domain, [self.form_id])
